Messy

Make a data frame messier.

Usage

messy(data, messiness = 0.1, missing = NA, case_type = "word")

Arguments

data

input dataframe

messiness

Percentage of values to change per function. Must be between 0 and 1. Default 0.1.

missing

A single value, vector, or list of what the missing values will be replaced with. If length is greater than 1, values will be replaced randomly. Default NA.

case_type

Whether the case should change based on the "word" or "letter".

Value

a dataframe the same size as the input data.

Examples

messy(mtcars)
#>                       mpg  cyl   disp   hp  drat     wt  qsec   vs   am gear
#> Mazda RX4              21    6   160  <NA>   3.9   2.62 16.46    0    1    4
#> Mazda RX4 Wag          21    6    160  110   3.9  2.875 17.02    0    1 <NA>
#> Datsun 710           22.8    4    108   93  <NA>   2.32 18.61   1     1    4
#> Hornet 4 Drive       21.4 <NA>    258  110 3.08  3.215  19.44    1    0    3
#> Hornet Sportabout    18.7    8    360  175  3.15   3.44 17.02    0    0    3
#> Valiant              18.1    6    225 105  2.76    3.46 20.22    1    0    3
#> Duster 360           14.3   8     360  245  3.21   3.57 15.84    0    0    3
#> Merc 240D            24.4    4 146.7  <NA>  <NA>   3.19   20  <NA>    0    4
#> Merc 230             22.8    4 140.8    95  3.92   3.15  22.9    1    0    4
#> Merc 280             19.2    6  167.6  123  3.92   3.44 18.3     1    0    4
#> Merc 280C            17.8 <NA>  167.6  123 3.92    <NA> 18.9  <NA>    0    4
#> Merc 450SE           16.4    8  275.8 180   <NA>   4.07  17.4    0    0    3
#> Merc 450SL           17.3    8  275.8  180  3.07   3.73  17.6    0    0    3
#> Merc 450SLC          15.2   8   275.8 <NA>  3.07   3.78    18    0    0    3
#> Cadillac Fleetwood   <NA>    8    472  205  2.93   5.25 17.98    0    0    3
#> Lincoln Continental  10.4    8    460  215     3  5.424 17.82   0     0    3
#> Chrysler Imperial    14.7    8    440  230 3.23  5.345  17.42    0    0    3
#> Fiat 128             32.4    4   78.7   66 4.08     2.2 19.47    1    1    4
#> Honda Civic          30.4    4  75.7   52   4.93  1.615  <NA>    1    1    4
#> Toyota Corolla       33.9    4   71.1   65  4.22  1.835 19.9     1    1 <NA>
#> Toyota Corona        <NA>    4  120.1   97   3.7  2.465 20.01    1 <NA>    3
#> Dodge Challenger     15.5    8    318  150  2.76   3.52 16.87    0    0    3
#> AMC Javelin          15.2    8    304  150  3.15  3.435  17.3    0    0    3
#> Camaro Z28           13.3    8   350   245  3.73   3.84 15.41    0    0 <NA>
#> Pontiac Firebird    19.2     8    400 <NA>  3.08  3.845 17.05    0    0   3 
#> Fiat X1-9            <NA>    4     79   66  4.08   <NA>  18.9    1 <NA>    4
#> Porsche 914-2         26     4  120.3 <NA>  4.43   2.14 16.7     0    1    5
#> Lotus Europa         30.4    4   95.1  113  3.77  1.513  16.9 <NA>    1    5
#> Ford Pantera L       15.8    8    351  264  4.22   3.17  <NA> <NA>    1 <NA>
#> Ferrari Dino         <NA>   6     145  175  3.62   2.77  15.5    0 <NA>    5
#> Maserati Bora          15   8    <NA>  335  3.54   3.57  14.6 <NA>    1    5
#> Volvo 142E           21.4    4    121 <NA> 4.11    2.78  18.6 <NA>    1    4
#>                     carb
#> Mazda RX4              4
#> Mazda RX4 Wag          4
#> Datsun 710             1
#> Hornet 4 Drive      <NA>
#> Hornet Sportabout      2
#> Valiant                1
#> Duster 360             4
#> Merc 240D              2
#> Merc 230               2
#> Merc 280            <NA>
#> Merc 280C              4
#> Merc 450SE             3
#> Merc 450SL             3
#> Merc 450SLC         <NA>
#> Cadillac Fleetwood     4
#> Lincoln Continental    4
#> Chrysler Imperial      4
#> Fiat 128               1
#> Honda Civic            2
#> Toyota Corolla         1
#> Toyota Corona          1
#> Dodge Challenger       2
#> AMC Javelin            2
#> Camaro Z28             4
#> Pontiac Firebird       2
#> Fiat X1-9              1
#> Porsche 914-2          2
#> Lotus Europa           2
#> Ford Pantera L         4
#> Ferrari Dino           6
#> Maserati Bora       <NA>
#> Volvo 142E             2