Change the timezone of datetime columns

Takes any number of datetime columns and changes their timezones either totally at random, or from a user-provided list of timezones.

Usage

messy_datetime_tzones(data, cols = NULL, tzones = OlsonNames(), force = FALSE)

Arguments

data

input dataframe

cols

set of columns to apply transformation to. If NULL will apply to all POSIXt columns.

tzones

Valid time zones to sample from. By default samples from all OlsonNames(), but can be set to options more relevant to the data.

force

By default (force = FALSE) the datetimes will have their actual hour/minute values changed along with the timezones. If force = TRUE, which requires lubridate, the datetime values will remain the same and only the timezone will differ.

Value

a dataframe the same size as the input data.

See also

Author

Jack Davison

Examples

data <- data.frame(dates = rep(Sys.time(), 10))

data$dates
#>  [1] "2025-08-15 22:11:30 UTC" "2025-08-15 22:11:30 UTC"
#>  [3] "2025-08-15 22:11:30 UTC" "2025-08-15 22:11:30 UTC"
#>  [5] "2025-08-15 22:11:30 UTC" "2025-08-15 22:11:30 UTC"
#>  [7] "2025-08-15 22:11:30 UTC" "2025-08-15 22:11:30 UTC"
#>  [9] "2025-08-15 22:11:30 UTC" "2025-08-15 22:11:30 UTC"
attr(data$dates, "tzone")
#> NULL

messy <- messy_datetime_tzones(data, tzones = "Poland")
messy$dates
#>  [1] "2025-08-16 00:11:30 CEST" "2025-08-16 00:11:30 CEST"
#>  [3] "2025-08-16 00:11:30 CEST" "2025-08-16 00:11:30 CEST"
#>  [5] "2025-08-16 00:11:30 CEST" "2025-08-16 00:11:30 CEST"
#>  [7] "2025-08-16 00:11:30 CEST" "2025-08-16 00:11:30 CEST"
#>  [9] "2025-08-16 00:11:30 CEST" "2025-08-16 00:11:30 CEST"
attr(messy$dates, "tzone")
#> [1] "Poland"